Change groupId to `io.github.pixelclover.uview`
