+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.uview</groupId>
+  <groupId>io.github.pixelclover.uview</groupId>
   <artifactId>uview</artifactId>
   <version>0.1.0-b.2</version>
   <packaging>jar</packaging>
@@ -161,7 +161,7 @@
               <createDependencyReducedPom>false</createDependencyReducedPom>
               <transformers>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.uview.App</mainClass>
+                  <mainClass>io.github.pixelclover.uview.App</mainClass>
                 </transformer>
               </transformers>
               <filters>
@@ -187,7 +187,7 @@
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>exec-maven-plugin</artifactId>
         <configuration>
-          <mainClass>com.uview.App</mainClass>
+          <mainClass>io.github.pixelclover.uview.App</mainClass>
         </configuration>
       </plugin>
       <plugin>
